lambdas/Aerial-Classifier/requirements.txt
tensorflow==2.4.1
requests~=2.31.0
pillow~=10.1.0
numpy~=1.19.2